requirement.txt
absl-py                   2.1.0
aiohttp                   3.9.3
aiosignal                 1.3.1
async-timeout             4.0.3
attrs                     23.2.0
certifi                   2024.2.2
charset-normalizer        3.3.2
click                     8.1.7
click-plugins             1.1.1
cligj                     0.7.2
cloudpickle               3.0.0
contourpy                 1.2.0
cycler                    0.12.1
distlib                   0.3.8
dm-tree                   0.1.8
filelock                  3.13.1
fiona                     1.9.5
fonttools                 4.49.0
frozenlist                1.4.1
fsspec                    2024.2.0
geopandas                 0.12.2
grpcio                    1.60.1
gym                       0.23.1
gym-notices               0.0.8
idna                      3.6
imageio                   2.34.0
importlib-metadata        7.0.1
importlib-resources       6.1.1
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
kiwisolver                1.4.5
lazy_loader               0.3
leuvenmapmatching         0.5.3
lightning-utilities       0.10.1
lz4                       4.3.3
Markdown                  3.5.2
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.8.3
mdurl                     0.1.2
msgpack                   1.0.7
multidict                 6.0.5
networkx                  3.2.1
numpy                     1.26.4
packaging                 23.2
pandas                    2.2.0
pillow                    10.2.0
pip                       23.3.1
platformdirs              4.2.0
protobuf                  4.25.3
pyDeprecate               0.3.2
Pygments                  2.17.2
pyparsing                 3.1.1
pyproj                    3.4.1
python-dateutil           2.8.2
pytorch-lightning         1.6.3
pytz                      2024.1
PyYAML                    6.0.1
ray                       2.2.0
referencing               0.33.0
requests                  2.31.0
rich                      13.7.0
rpds-py                   0.18.0
scikit-image              0.22.0
scipy                     1.12.0
setuptools                68.2.2
shapely                   2.0.3
six                       1.16.0
sumolib                   1.16.0
svgwrite                  1.4.3
tabulate                  0.9.0
tensorboard               2.16.2
tensorboard-data-server   0.7.2
tensorboardX              2.6.2.2
tifffile                  2024.2.12
torch                     1.12.1+cu113
torch-scatter             2.1.2
torchaudio                0.12.1+cu113
torchmetrics              1.3.1
torchvision               0.13.1+cu113
tqdm                      4.66.2
Tree                      0.2.4
typer                     0.9.0
typing_extensions         4.9.0
tzdata                    2024.1
urllib3                   2.2.1
virtualenv                20.25.0
Werkzeug                  3.0.1
wheel                     0.41.2
yarl                      1.9.4
zipp                      3.17.0
